% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age-categories.R
\docType{data}
\name{age_categories}
\alias{age_categories}
\title{Prices of 50,000 round cut diamonds.}
\format{A data frame with 53940 rows and 10 variables:}
\source{
\url{http://www.diamondse.info/}
}
\usage{
age_categories
}
\description{
A dataset containing the prices and other attributes of almost 54,000
diamonds.
}
\keyword{datasets}
