% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cases}
\alias{cases}
\title{cases}
\format{
Dataframe with 5 columns, number of rows will differ per update.
}
\source{
\url{https://www.ecdc.europa.eu/en/covid-19/data}
}
\usage{
cases
}
\description{
A dataframe with 5 variables: date (Date), cases (num), deaths (num),
country (chr) and cumulative_death (num).
It records per country and per date, the number of COVID
cases and the number of deaths, both on the day and cumulative.
}
\keyword{datasets}
