% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{contact_home}
\alias{contact_home}
\title{contact_home}
\format{
A Large list of 152 tibbles, each a 16x16 list of doubles.
}
\source{
\url{https://journals.plos.org/ploscompbiol/article?id=10.1371/journal.pcbi.1005697#sec020}
}
\usage{
contact_home
}
\description{
A list of 152 16x16 asymmetrical contact matrices, one for each country,
describing which age groups have contact with which other age groups AT HOME.
The units of entries are the average number of contacts with a person in
that age group in a day. The top row is individuals in the age group 0-5
years, the second row is individuals in the age group 5-10 years, and so
on torow 16 which is individuals aged 75-80. The columns represent the
people they come into contact with, where column 1 is the age groups 0-5
(X1), column to is the age group 5-10 (X2), and so to column 16 (ages 75-80).
}
\keyword{datasets}
