% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contact-school.R
\docType{data}
\name{contact_school}
\alias{contact_school}
\title{Prices of 50,000 round cut diamonds.}
\format{A data frame with 53940 rows and 10 variables:
\describe{
\item{price}{price, in US dollars}
\item{carat}{weight of the diamond, in carats}
...
}}
\source{
\url{http://www.diamondse.info/}
}
\usage{
contact_school
}
\description{
A dataset containing the prices and other attributes of almost 54,000
diamonds.
}
\keyword{datasets}
