% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{countries_demog}
\alias{countries_demog}
\title{countries_demog}
\format{
list of chr
}
\usage{
countries_demog
}
\description{
A list of 201 countries (chr) that the model can take, in alphabetical order.
}
\keyword{datasets}
