% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateExcelFun.R
\name{create_excel}
\alias{create_excel}
\title{Creates an excel sheet with country-specific data and returns the name of
the excel file. The user should input the country of choice. If they do not,
or the country name does not appear in countries_demog, the UK is used.
The data_CoMo.Rdata should be loaded before this function is called. This
function also relies on the package XLConnect, which requires Java. It
takes the /data/Template_CoMo_CountryData_temp.xlsx file, copies it and
fills in the Cases and Population sheets with country-specific data.}
\usage{
create_excel(country)
}
\arguments{
\item{country}{string of characters, optional, the country for which the
data should be generated}
}
\value{
name of the excel file that contains the country specific data.
}
\description{
Creates an excel sheet with country-specific data and returns the name of
the excel file. The user should input the country of choice. If they do not,
or the country name does not appear in countries_demog, the UK is used.
The data_CoMo.Rdata should be loaded before this function is called. This
function also relies on the package XLConnect, which requires Java. It
takes the /data/Template_CoMo_CountryData_temp.xlsx file, copies it and
fills in the Cases and Population sheets with country-specific data.
}
