% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{population}
\alias{population}
\title{population}
\format{
Dataframe with 5 columns and 4221 rows.
}
\source{
\url{https://population.un.org/wpp/Download/Standard/Population/}
}
\usage{
population
}
\description{
A dataframe with 5 variables: country (chr), age_category (Factor),
pop (num), birth (num), death(num). It specifies per country and per age
category, the size of the population, the number of births per person per
day and the number of deaths per person per day. The age categories are
0-4 years, 5-9 years, 10-14 years, and so on, up to the last category,
which is 100+ years old.
}
\keyword{datasets}
