% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cases.R
\docType{data}
\name{cases}
\alias{cases}
\title{Prices of 50,000 round cut diamonds.}
\format{
A data frame with 53940 rows and 10 variables:
}
\source{
\url{http://www.diamondse.info/}
}
\usage{
cases
}
\description{
A dataset containing the prices and other attributes of almost 54,000
diamonds.
}
\keyword{datasets}
