% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countries-demog.R
\docType{data}
\name{countries_demog}
\alias{countries_demog}
\title{Prices of 50,000 round cut diamonds.}
\format{A data frame with 53940 rows and 10 variables:}
\source{
\url{http://www.diamondse.info/}
}
\usage{
countries_demog
}
\description{
A dataset containing the prices and other attributes of almost 54,000
diamonds.
}
\keyword{datasets}
