% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_contact_matrix_try.R
\name{get_contact_matrix}
\alias{get_contact_matrix}
\title{Create a contact matrix}
\usage{
get_contact_matrix(country, contact_type, n_5yr_agegroups_popstruc)
}
\arguments{
\item{country}{String representing the country for which the contact matrix will
be loaded. Naming conventions for countries must be in line with those used
in input excel file.}

\item{contact_type}{A loaded matrix representing the contact between different age groups
in a given social setting e.g. contact_work gives contact matrix for the work place.}

\item{n_5yr_agegroups_popstruc}{Number of 5 year age groups starting with 0-4yr in the population.}
}
\value{
A square contact matrix for specified country with dimension equal to \code{n_5yr_agegroups_popstruc}
}
\description{
Function that loads and processes a contact matrix for use in the CoMo model ode solver
}
\examples{
get_contact_matrix("United Kingdom of Great Britain", contact_home, 21)

}
